chore: move FakeKeyring fixture to conftest
All unit test should use the mocked keyring.
